Add support for Xilinx HDMI and refactor video into submodule
