+++ pynq/lib/_pynq/_xhdmi/hdmi_functions.c
@@ -4,6 +4,43 @@
 #include <stdio.h>
 #include <libxlnk_cma.h>
 
+unsigned char edid_1920x1080[] = {
+	0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x00,
+	0x61, 0x98, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x01, 0x18, 0x01, 0x03, 0xa1, 0x00, 0x00, 0x80,
+	0x06, 0xee, 0x91, 0xa3, 0x54, 0x4c, 0x99, 0x26,
+	0x0f, 0x50, 0x54, 0x21, 0x80, 0x00, 0xd1, 0x00,
+	0xd1, 0xc0, 0x81, 0x00, 0x81, 0xc0, 0x01, 0x01,
+	0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x02, 0x3a,
+	0x80, 0x18, 0x71, 0x38, 0x2d, 0x40, 0x58, 0x2c,
+	0x45, 0x00, 0x55, 0x50, 0x21, 0x00, 0x00, 0x1e,
+	0x08, 0xe8, 0x00, 0x30, 0xf2, 0x70, 0x5a, 0x80,
+	0xb0, 0x58, 0x8a, 0x00, 0x55, 0x50, 0x21, 0x00,
+	0x00, 0x1e, 0x00, 0x00, 0x00, 0xfc, 0x00, 0x58,
+	0x69, 0x6c, 0x69, 0x6e, 0x78, 0x20, 0x50, 0x59,
+	0x4e, 0x51, 0x0a, 0x20, 0x00, 0x00, 0x00, 0xfd,
+	0x00, 0x1d, 0x56, 0x1e, 0x8c, 0x3c, 0x00, 0x0a,
+	0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x01, 0x34,
+	0x02, 0x03, 0x42, 0xf1, 0x51, 0x61, 0x60, 0x5f,
+	0x5e, 0x5d, 0x10, 0x1f, 0x20, 0x05, 0x14, 0x04,
+	0x13, 0x12, 0x11, 0x03, 0x02, 0x01, 0x23, 0x09,
+	0x1f, 0x07, 0x83, 0x01, 0x00, 0x00, 0x6d, 0x03,
+	0x0c, 0x00, 0x10, 0x00, 0x38, 0x3c, 0x20, 0x00,
+	0x60, 0x01, 0x02, 0x03, 0x67, 0xd8, 0x5d, 0xc4,
+	0x01, 0x78, 0x80, 0x03, 0xe2, 0x0f, 0x03, 0xe3,
+	0x05, 0xff, 0x01, 0xe6, 0x06, 0x07, 0x01, 0x8b,
+	0x60, 0x11, 0x56, 0x5e, 0x00, 0xa0, 0xa0, 0xa0,
+	0x29, 0x50, 0x30, 0x20, 0x35, 0x00, 0x55, 0x50,
+	0x21, 0x00, 0x00, 0x1a, 0x11, 0x44, 0x00, 0xa0,
+	0x80, 0x00, 0x1f, 0x50, 0x30, 0x20, 0x36, 0x00,
+	0x55, 0x50, 0x21, 0x00, 0x00, 0x1a, 0xbf, 0x16,
+	0x00, 0xa0, 0x80, 0x38, 0x13, 0x40, 0x30, 0x20,
+	0x3a, 0x00, 0x55, 0x50, 0x21, 0x00, 0x00, 0x1a,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xca
+};
+
+unsigned int edid_1920x1080_len = 256;
+
 typedef struct {
 	XVphy Vphy;
 	XVphy_Config config;
@@ -63,6 +100,14 @@ void VphyHdmiRxReadyCallback(void *CallbackRef) {
         }
 }
 
+void TxVsCallback(void *CallbackRef) {
+	HdmiTx* tx = (HdmiTx*)CallbackRef;
+	XHdmiC_AVI_InfoFrame *AviInfoFramePtr;
+	AviInfoFramePtr = XV_HdmiTxSs_GetAviInfoframe(&tx->HdmiTxSs);
+	XHdmiC_Aux info_frame = XV_HdmiC_AVIIF_GeneratePacket(AviInfoFramePtr);
+	XV_HdmiTxSs_SendGenericAuxInfoframe(&tx->HdmiTxSs, &info_frame);
+}
+
 void* HdmiPhy_new(unsigned long BaseAddress) {
 	HdmiPhy* phy = (HdmiPhy*)calloc(1, sizeof(HdmiPhy));
 	if (!phy) return 0;
@@ -86,6 +131,11 @@ void HdmiPhy_handle_events(void* handle) {
 	XVphy_InterruptHandler(&phy->Vphy);
 }
 
+void HdmiPhy_report(void* handle) {
+	HdmiPhy* phy = (HdmiPhy*)handle;
+	XVphy_HdmiDebugInfo(&phy->Vphy, 0, XVPHY_CHANNEL_ID_CH1);
+}
+
 void* HdmiRx_new(unsigned long BaseAddress, void* phy_handle) {
 	HdmiRx* rx = (HdmiRx*)calloc(1, sizeof(HdmiRx));
 	if (!rx) return 0;
@@ -99,6 +149,8 @@ void* HdmiRx_new(unsigned long BaseAddress, void* phy_handle) {
 		free(rx);
 		return 0;
 	}
+	XV_HdmiRxSs_SetEdidParam(&rx->HdmiRxSs, edid_1920x1080, edid_1920x1080_len);
+	XV_HdmiRxSs_LoadDefaultEdid(&rx->HdmiRxSs);
 	rx->phy = (HdmiPhy*)phy_handle;
         XVphy_SetHdmiCallback(&rx->phy->Vphy,
 		XVPHY_HDMI_HANDLER_RXREADY,
@@ -110,6 +162,7 @@ void* HdmiRx_new(unsigned long BaseAddress, void* phy_handle) {
 		(void *)rx);
 	XV_HdmiRxSs_SetCallback(&rx->HdmiRxSs, XV_HDMIRXSS_HANDLER_STREAM_INIT,
 		(void *)RxStreamInitCallback, (void *)rx);
+	XV_HdmiRxSs_SetHpd(&rx->HdmiRxSs, FALSE);
 	return rx;
 }
 
@@ -145,6 +198,27 @@ int HdmiRx_vsize(void* handle) {
 	return stream->Timing.VActive;
 }
 
+int HdmiRx_fps(void* handle) {
+	HdmiRx* rx = (HdmiRx*)handle;
+	XVidC_VideoStream *stream = XV_HdmiRxSs_GetVideoStream(&rx->HdmiRxSs);
+	return stream->FrameRate;
+}
+
+void HdmiRx_report(void* handle) {
+	HdmiRx* rx = (HdmiRx*)handle;
+	XV_HdmiRxSs_ReportInfo(&rx->HdmiRxSs);
+}
+
+void HdmiRx_load_edid(void* handle, unsigned char* data, unsigned length) {
+	HdmiRx* rx = (HdmiRx*)handle;
+	XV_HdmiRxSs_LoadEdid(&rx->HdmiRxSs, data, length);
+}
+
+void HdmiRx_set_hpd(void* handle, unsigned value) {
+	HdmiRx* rx = (HdmiRx*)handle;
+	XV_HdmiRxSs_SetHpd(&rx->HdmiRxSs, value);
+}
+
 void* HdmiTx_new(unsigned long BaseAddress, void* phy_handle) {
 	HdmiTx* tx = (HdmiTx*)calloc(1, sizeof(HdmiTx));
 	if (!tx) return 0;
@@ -166,6 +240,10 @@ void* HdmiTx_new(unsigned long BaseAddress, void* phy_handle) {
 		(void *)VphyHdmiTxInitCallback,
 		(void *)tx);
 
+	XV_HdmiTxSs_SetCallback(&tx->HdmiTxSs,
+			XV_HDMITXSS_HANDLER_VS,
+			(void*) TxVsCallback,
+			(void*) tx);
 	return tx;
 }
 
@@ -184,29 +262,81 @@ int HdmiTx_connected(void* handle) {
 	return XV_HdmiTxSs_IsStreamConnected(&tx->HdmiTxSs);
 }
 
-int HdmiTx_set_format(void* handle, int hsize, int vsize) {
-	XVidC_VideoMode mode = XVidC_GetVideoModeId(hsize, vsize, 60, 0);
-	if (mode == XVIDC_VM_NOT_SUPPORTED) return 0;
+int HdmiTx_set_format(void* handle, int hsize, int vsize, int fps) {
+	XVidC_VideoMode mode = XVidC_GetVideoModeId(hsize, vsize, fps, 0);
+	XHdmiC_AVI_InfoFrame *AviInfoFramePtr;
+	if (mode == XVIDC_VM_NOT_SUPPORTED) return -1;
 	
 	HdmiTx* tx = (HdmiTx*)handle;
+	XV_HdmiTxSs_DetectHdmi20(&tx->HdmiTxSs);
+	AviInfoFramePtr = XV_HdmiTxSs_GetAviInfoframe(&tx->HdmiTxSs);
 	int clock = XV_HdmiTxSs_SetStream(&tx->HdmiTxSs, mode, XVIDC_CSF_RGB, XVIDC_BPC_8, NULL);
+
 	tx->phy->Vphy.HdmiTxRefClkHz = clock;
-	return clock;
+	int status = XVphy_SetHdmiTxParam(&tx->phy->Vphy, 0, XVPHY_CHANNEL_ID_CHA, 2, 8, XVIDC_CSF_RGB);
+	if (status != XST_SUCCESS) {
+		return -2;
+	}
+	AviInfoFramePtr->Version = 2;
+	AviInfoFramePtr->ColorSpace = XV_HdmiC_XVidC_To_IfColorformat(XVIDC_CSF_RGB);
+	AviInfoFramePtr->VIC = tx->HdmiTxSs.HdmiTxPtr->Stream.Vic;
+	return tx->phy->Vphy.HdmiTxRefClkHz;
+}
+
+unsigned long long HdmiTx_line_rate(void* handle) {
+	HdmiTx* tx = (HdmiTx*)handle;
+	XVphy_PllType TxPllType;
+	TxPllType = XVphy_GetPllType(&tx->phy->Vphy, 0, XVPHY_DIR_TX, XVPHY_CHANNEL_ID_CH1);
+	if ((TxPllType == XVPHY_PLL_TYPE_CPLL)) {
+		return XVphy_GetLineRateHz(&tx->phy->Vphy, 0, XVPHY_CHANNEL_ID_CH1);
+	} else if((TxPllType == XVPHY_PLL_TYPE_QPLL) ||
+			  (TxPllType == XVPHY_PLL_TYPE_QPLL0) ||
+			  (TxPllType == XVPHY_PLL_TYPE_PLL0)) {
+		return XVphy_GetLineRateHz(&tx->phy->Vphy, 0, XVPHY_CHANNEL_ID_CMN0);
+	} else {
+		return XVphy_GetLineRateHz(&tx->phy->Vphy, 0, XVPHY_CHANNEL_ID_CMN1);
+	}
+
 }
 
 int HdmiTx_start(void* handle) {
 	HdmiTx* tx = (HdmiTx*)handle;
-	int status = XVphy_SetHdmiTxParam(&tx->phy->Vphy, 0, XVPHY_CHANNEL_ID_CHA, 2, 8, XVIDC_CSF_RGB);
-	if (status != XST_SUCCESS) {
-		return 0;
-	}
 	XVphy_Clkout1OBufTdsEnable(&tx->phy->Vphy, XVPHY_DIR_TX, (TRUE));
 	XVphy_IBufDsEnable(&tx->phy->Vphy, 0, XVPHY_DIR_TX, (TRUE));
+	XV_HdmiTxSs_SetSamplingRate(&tx->HdmiTxSs, tx->phy->Vphy.HdmiTxSampleRate);
 	XV_HdmiTxSs_StreamStart(&tx->HdmiTxSs);
+	XV_HdmiTxSs_AudioMute(&tx->HdmiTxSs, TRUE);
+	return 1;
 }
+
+void HdmiTx_stop(void* handle) {
+	HdmiTx* tx = (HdmiTx*)handle;
+	XVphy_Clkout1OBufTdsEnable(&tx->phy->Vphy, XVPHY_DIR_TX, (FALSE));
+}
+
+void HdmiTx_dvi_mode(void* handle) {
+	HdmiTx* tx = (HdmiTx*)handle;
+	XV_HdmiTxSS_SetDviMode(&tx->HdmiTxSs);
+}
+
+void HdmiTx_hdmi_mode(void* handle) {
+	HdmiTx* tx = (HdmiTx*)handle;
+	XV_HdmiTxSS_SetHdmiMode(&tx->HdmiTxSs);
+}
+
+int HdmiTx_read_edid(void* handle, unsigned char* data) {
+	HdmiTx* tx = (HdmiTx*)handle;
+	return XV_HdmiTxSs_ReadEdid(&tx->HdmiTxSs, data);
+}
+
 
 int HdmiTx_ready(void* handle) {
 	HdmiTx* tx = (HdmiTx*)handle;
 	return XV_HdmiTxSs_IsStreamUp(&tx->HdmiTxSs);
+}
+
+void HdmiTx_report(void* handle) {
+	HdmiTx* tx = (HdmiTx*)handle;
+	XV_HdmiTxSs_ReportInfo(&tx->HdmiTxSs);
 }
